requirements.txt
django
djangorestframework
gunicorn
pytz